requirements.txt
flask 
scikit-learn 
pandas 
numpy 
tensorflow 
pillow
matplotlib